+++ src/packet_sources/tun.rs
@@ -1,5 +1,4 @@
 use anyhow::{Context, Result};
-
 use crate::messages::{
     NetworkCommand, NetworkEvent, SmolPacket, TransportCommand, TransportEvent, TunnelInfo,
 };
@@ -90,7 +89,6 @@ impl PacketSourceTask for TunTask {
                         log::error!("Skipping invalid packet from tun interface: {:?}", &buf[..len]);
                         continue;
                     };
-                    dbg!(&packet);
                     permit.take().unwrap().send(NetworkEvent::ReceivePacket {
                         packet,
                         tunnel_info: TunnelInfo::None,
@@ -98,13 +96,15 @@ impl PacketSourceTask for TunTask {
                 },
                 // send_to is cancel safe, so we can use that for backpressure.
                 r = self.device.send(&packet_to_send), if !packet_to_send.is_empty() => {
-                    r.context("TUN write() failed")?;
+                    let sent = r.context("TUN write() failed")?;
+                    if sent != packet_to_send.len() {
+                        log::debug!("device.send: {} of {} bytes sent.", sent, packet_to_send.len());
+                    }
                     packet_to_send.clear();
                 },
                 Some(command) = self.net_rx.recv(), if packet_to_send.is_empty() => {
                     match command {
                         NetworkCommand::SendPacket(packet) => {
-                            dbg!(&packet);
                             packet_to_send = packet.into_inner();
                         }
                     }